BoringSSL tests: align with new structure
- Move TestState to separate file
- Move RetryAsync to handshake_util
- Move Get/Set TestConfig to test_config
- Move SetupCTX and NewSSL to test_config (including all callbacks)
