+++ test/ossl_shim/build.info
@@ -1,6 +1,6 @@
 IF[{- defined $target{CXX} && !$disabled{"external-tests"} -}]
   PROGRAMS{noinst}=ossl_shim
-  SOURCE[ossl_shim]=ossl_shim.cc async_bio.cc packeted_bio.cc test_config.cc
+  SOURCE[ossl_shim]=ossl_shim.cc async_bio.cc handshake_util.cc packeted_bio.cc test_config.cc test_state.cc
   INCLUDE[ossl_shim]=. include ../../include
   DEPEND[ossl_shim]=../../libssl ../../libcrypto
 ENDIF
